CHORE: Adjusted endpoints to be consistent

+++ src/Web/Actions/Gallery/Positions/DeletePositionAction.php
@@ -6,6 +6,7 @@ use App\Database\GalleryFilePosition;
 use App\Web\Actions\Action;
 use JsonException;
 use Psr\Http\Message\ResponseInterface as Response;
+use Slim\Exception\HttpNotFoundException;
 
 class DeletePositionAction extends Action
 {
@@ -13,11 +14,16 @@ class DeletePositionAction extends Action
     /**
      * @inheritDoc
      * @throws JsonException
+     * @throws HttpNotFoundException
      */
     protected function action(): Response
     {
-        $positionId = $this->args['galleryFileId'];
-        $galleryFilePosition = GalleryFilePosition::findById($positionId);
+        $galleryId = $this->args['galleryId'];
+        $position = $this->args['position'];
+        $galleryFilePosition = GalleryFilePosition::findByPosition($galleryId, $position);
+        if (!$galleryFilePosition) {
+            throw new HttpNotFoundException($this->request, 'Gallery file position not found');
+        }
         $galleryFilePosition->delete();
 
         return $this->noContent();
